remove option to set verbose flag in this package, more doc
The ganglia package can use vlog but should not have an option to
set the flag: this flag should be set externally by anything
that actually wants to use vlog.

Combines SetPrefix and Configure of the groupname into Configure
for both options.

Adds a bit of doc describing vlog / Configure