strftime: break build into linux/non-linux
And change the tests to follow.

In particular, strftime on OS X will yield slightly different results in
the face of invalid format specifiers.

Further, once you change a platform to use the Pure version, the
existing Perl tests will only complain about the fact that you are not
using the system version.