Return child errors to callers of CreateChild
Previously errors that occurred in the privsep child were written to the
childOut file descriptor and left to the user to sort out. This creates
a separate status pipe so errors can be noticed in the parent and
returned as an error from CreateChild.

Also add a specific error for the case where the running binary is not
executable by the unprivileged user.

On that note, add a comment for why `sudo go test` doesn't work.